Refactor the cluster detail page components to be more flexible
