--- src/components/UI/Display/MAPI/releases/types.ts
@@ -1,4 +1,0 @@
-export interface IRelease {
-  version: string;
-  k8sVersion?: string;
-}
